[add] calculate price method
