--- src/main/java/service/CarService.java
@@ -1,22 +1,0 @@
-package service;
-
-import model.Car;
-import model.CarType;
-
-import java.time.LocalDate;
-import java.util.List;
-
-public class CarService {
-    public List<Car> getFeaturedCars() {
-        return null;
-    }
-
-    public List<Car> getAvailableCarsByDate(LocalDate from, LocalDate to) {
-        return null;
-    }
-
-
-    public List<Car> getAvailableCarsByType(List<CarType> types) {
-        return null;
-    }
-}
